DBZ-3546 merge the two Pravega sinks into one sink
