+++ debezium-server/debezium-server-pravega/src/main/java/io/debezium/server/pravega/PravegaSink.java
@@ -1,0 +1,8 @@
+package io.debezium.server.pravega;
+
+import io.debezium.engine.ChangeEvent;
+import io.debezium.engine.DebeziumEngine.ChangeConsumer;
+
+public interface PravegaSink extends ChangeConsumer<ChangeEvent<Object, Object>>, AutoCloseable {
+
+}
