+++ debezium-server/debezium-server-pravega/src/test/java/io/debezium/server/pravega/PravegaTxnIT.java
@@ -36,7 +36,7 @@ import io.quarkus.test.junit.QuarkusTest;
 @QuarkusTest
 @QuarkusTestResource(PostgresTestResourceLifecycleManager.class)
 @QuarkusTestResource(PravegaTestResource.class)
-@EnabledIfSystemProperty(named = "debezium.sink.type", matches = "pravega-txn")
+@EnabledIfSystemProperty(named = "debezium.sink.pravega.transaction", matches = "true")
 public class PravegaTxnIT {
 
     private static final int MESSAGE_COUNT = 4;
